refactor: uistate 모델 각 패키지로 이동
