+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/home/ticketlist/TicketListAdapter.kt
@@ -3,7 +3,6 @@ package com.festago.festago.presentation.ui.home.ticketlist
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import com.festago.festago.presentation.model.TicketListItemUiState
 
 class TicketListAdapter :
     ListAdapter<TicketListItemUiState, TicketListItemViewHolder>(ticketDiffUtil) {
