+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketreserve/TicketReserveUiState.kt
@@ -1,8 +1,5 @@
 package com.festago.festago.presentation.ui.ticketreserve
 
-import com.festago.festago.presentation.model.ReservationFestivalUiState
-import com.festago.festago.presentation.model.TicketReserveItemUiState
-
 sealed interface TicketReserveUiState {
     object Loading : TicketReserveUiState
     class Success(
