+++ android/festago/app/src/main/java/com/festago/festago/presentation/ui/ticketreserve/adapter/TicketReserveAdapter.kt
@@ -3,7 +3,7 @@ package com.festago.festago.presentation.ui.ticketreserve.adapter
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import com.festago.festago.presentation.model.TicketReserveItemUiState
+import com.festago.festago.presentation.ui.ticketreserve.TicketReserveItemUiState
 import com.festago.festago.presentation.ui.ticketreserve.viewHolder.TicketReserveViewHolder
 
 class TicketReserveAdapter :
